feat: ✨ Refactor update and install subcomands
